Refact - UserProfile 분리
